split k3 model.py to moe.py lda.py
